+++ src/components/Organism/Resume/index.ts
@@ -1,4 +1,4 @@
-export * from './ResumeWorkExperience'
 export * from './ResumeExperience'
+export * from './ResumeActivities'
 export * from './ResumeIntro'
 export * from './ResumeSkills'
